chore(cxp): higiene + cierre honesto de Sprint 6 (rediseño confirmado en prod)
Revisión de pendientes contra prod. El rediseño del flujo en 3 etapas
(Sprint 7) ya estaba cerrado en prod (#1068/#1069); el doc lo reconcilia.

Higiene de código:
- Elimina `components/cxp/cxp-programacion-module.tsx` (~760 líneas): el
  componente de programación en lote dejó de renderizarse tras S7 (#1069).
- Extrae el helper puro `pendientesDeProgramar` + tipo `AplicacionViva`
  (vivos: bandeja «Te toca» + vista "Por programar" de Facturas) a
  `lib/cxp/pendientes-programar.ts` con su test. Reapunta los 3 importadores.

Cierre honesto de Sprint 6 (en el planning doc, por decisión no por código):
- `erp.gastos` vacía (0 filas) → "migración de gastos" es no-op. NO se dropea:
  `erp.conciliaciones` (legacy) la referencia por FK; su destino es de
  `conciliacion-bancaria`.
- COAGAN/ANSA sin datos ERP ni rol Dirección → rollout diferido (sin valor hoy).
- Bug latente `cxc_pago_registrar tipo='ingreso'` verificado ya corregido en
  prod (inserta `tipo='abono'`).

Único pendiente sustantivo restante = conciliación movimiento↔banco (Sprint 5),
que pertenece a la hermana `conciliacion-bancaria`.
